make rust code available in snakemake
