+++ peptonizer_rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "peptonizer"
+name = "peptonizer_rust"
 version = "0.1.0"
 publish = false
 edition = "2024"
@@ -10,8 +10,16 @@ crate-type = ["cdylib"]
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+csv = "1.1"
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
 web-sys = { version = "0.3", features = ['Headers', 'Request', 'RequestInit', 'RequestMode', 'Response', 'Window', 'XmlHttpRequest'] }
 js-sys = "0.3.76"
-csv = "1.1"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+pyo3 = { version = "0.18", features = ["extension-module"] }
+reqwest = { version = "0.11", features = ["json"] }
+tokio = { version = "1", features = ["full"] }
+rand = "0.8"
