+++ peptonizer_rust/src/random.rs
@@ -1,11 +1,13 @@
+#[cfg(target_arch = "wasm32")]
 use js_sys::Math;
+
 use std::collections::HashSet;
 
 #[cfg(target_arch = "wasm32")]
 pub fn select_random_samples_with_weights(
     weights: Vec<f64>,
     n: usize,
-) -> Vec<usize> {
+) -> HashSet<usize> {
 
     let cumulative_weights: Vec<f64> = weights
         .iter()
@@ -15,8 +17,7 @@ pub fn select_random_samples_with_weights(
         })
         .collect();
 
-    let mut samples: Vec<usize> = Vec::with_capacity(n);
-
+    let mut samples: HashSet<usize> = HashSet::with_capacity(n);
 
     while samples.len() < n {
         let r = Math::random() as f64; 
@@ -26,11 +27,30 @@ pub fn select_random_samples_with_weights(
 
         let chosen_idx: usize = cumulative_weights.binary_search_by(|&w| w.partial_cmp(&r).unwrap()).unwrap_or_else(|x| x);
 
-        samples.push(chosen_idx);
+        samples.insert(chosen_idx);
     }
 
-    let samples_set: HashSet<usize> = samples.iter().cloned().collect();
-    let samples: Vec<usize> = samples_set.iter().cloned().collect();
-
     samples
+}
+
+#[cfg(not(target_arch = "wasm32"))]
+pub fn select_random_samples_with_weights(
+    weights: Vec<f64>,
+    n: usize,
+) -> HashSet<usize> {
+    use rand::prelude::*;
+    use rand::distributions::WeightedIndex;
+
+    // Create a WeightedIndex for sampling
+    let dist = WeightedIndex::new(&weights).unwrap();
+
+    let mut rng = thread_rng();
+    let mut selected_indices = HashSet::new();
+
+    while selected_indices.len() < n {
+        let idx = dist.sample(&mut rng);
+        selected_indices.insert(idx);
+    }
+
+    selected_indices
 }
