+++ peptonizer_rust/src/utils.rs
@@ -1,3 +1,4 @@
+#[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
 use serde::Deserialize;
