+++ snakemake/workflow/Snakefile
@@ -39,6 +39,7 @@ rule GetTaxonomyFromUnipept:
     shell:
         """
         export PYTHONPATH={project_root}:${{PYTHONPATH:-}}
+        pip install --force-reinstall wheels/peptonizer_rust-0.1.0-cp312-cp312-manylinux_2_34_aarch64.whl
         {params.run_command} scripts/pout_to_taxonomy.py --input-file {input[0]} --unipept-response-file {output[0]} --log-file {log[0]} --taxonomy-query {params.taxon_query} --taxon-rank {params.taxon_rank}
         """
 
@@ -59,6 +60,7 @@ rule WeightTaxa:
     shell:
         """
         export PYTHONPATH={project_root}:${{PYTHONPATH:-}}
+        pip install --force-reinstall wheels/peptonizer_rust-0.1.0-cp312-cp312-manylinux_2_34_aarch64.whl
         {params.run_command} scripts/infer_taxa_weights.py --pout-file {input[0]} --unipept-response-file {input[1]}  --sequence-scores-dataframe-file {output[0]} --taxa-weights-dataframe-file {output[1]} --number-of-taxa {params.number_of_taxa} --taxon-rank {params.taxon_rank}
         """
