+++ snakemake/workflow/scripts/infer_taxa_weights.py
@@ -3,6 +3,7 @@ import gzip
 import json
 
 from peptonizer.peptonizer import perform_taxa_weighing, parse_peptide_tsv
+from peptonizer_rust import perform_taxa_weighing_py
 
 parser = argparse.ArgumentParser()
 
@@ -54,18 +55,20 @@ with open(args.pout_file, 'rt', encoding='utf-8') as file:
 pep_score, pep_psm_counts = parse_peptide_tsv(file_contents)
 
 
-# Parse the Unipept response file
+# Read the Unipept response file
+unipept_responses = ""
 with open(args.unipept_response_file, "r") as file:
-    unipept_responses = json.load(file)
+    unipept_responses = file.read()
 
-df, weights = perform_taxa_weighing(
+sequence_scores, taxa_weights = perform_taxa_weighing_py(
     unipept_responses,
-    pep_score,
-    pep_psm_counts,
+    json.dumps(pep_score),
+    json.dumps(pep_psm_counts),
     args.number_of_taxa,
     args.taxon_rank
 )
 
 print("Started dumping produced results to CSV-files...")
-df.to_csv(args.sequence_scores_dataframe_file)
-weights.to_csv(args.taxa_weights_dataframe_file)
+with open(args.sequence_scores_dataframe_file, 'w') as sequences_file, open(args.taxa_weights_dataframe_file, 'w') as weights_file:
+    sequences_file.write(sequence_scores)
+    weights_file.write(taxa_weights)
